fix(app): extensions in available apps grid, only OpenAI Image Gen and Ollama
