+++ apps/app/src/react-app/shell/settings-route.tsx
@@ -131,7 +131,7 @@ import {
   openAiImageResponseToArrayBuffer,
   slugifyImageArtifactName,
 } from "../domains/settings/openai-image-extension";
-import type { LocalProviderInstallInput } from "../domains/settings/local-provider-extension-card";
+import { OLLAMA_PROVIDER_CONFIG, type LocalProviderInstallInput } from "../domains/settings/openai-image-extension";
 
 type RouteWorkspace = OpenworkWorkspaceInfo & {
   displayNameResolved: string;
@@ -1962,27 +1962,16 @@ function SettingsRouteContent() {
                   if (entry.kind === "plugin") {
                     const id = entry.serverName ?? entry.name;
                     if (id === "openai-image-gen") {
-                      // OpenAI Image Gen needs an API key; prompt handled by detail modal.
-                      // For now, just install with empty key (user configures in Settings -> Environment).
                       void installOpenAiImageExtension("");
-                    } else {
-                      // Local provider extensions (ollama, lmstudio, llama-cpp)
-                      const LOCAL_PROVIDERS: Record<string, { name: string; baseURL: string; defaultModelId: string }> = {
-                        ollama: { name: "Ollama (local)", baseURL: "http://localhost:11434/v1", defaultModelId: "qwen2.5-coder:7b" },
-                        lmstudio: { name: "LM Studio (local)", baseURL: "http://127.0.0.1:1234/v1", defaultModelId: "google/gemma-3n-e4b" },
-                        "llama-cpp": { name: "llama-server (local)", baseURL: "http://127.0.0.1:8080/v1", defaultModelId: "qwen3-coder:a3b" },
-                      };
-                      const preset = LOCAL_PROVIDERS[id];
-                      if (preset) {
-                        void installLocalProvider({
-                          providerId: id === "llama-cpp" ? "llama.cpp" : id,
-                          name: preset.name,
-                          baseURL: preset.baseURL,
-                          modelId: preset.defaultModelId,
-                          modelName: preset.defaultModelId,
-                          setDefault: true,
-                        });
-                      }
+                    } else if (id === "ollama") {
+                      void installLocalProvider({
+                        providerId: OLLAMA_PROVIDER_CONFIG.providerId,
+                        name: OLLAMA_PROVIDER_CONFIG.name,
+                        baseURL: OLLAMA_PROVIDER_CONFIG.baseURL,
+                        modelId: OLLAMA_PROVIDER_CONFIG.defaultModelId,
+                        modelName: OLLAMA_PROVIDER_CONFIG.defaultModelId,
+                        setDefault: true,
+                      });
                     }
                     return;
                   }
